src/azure_functions/requirements.txt
azure-functions>=1.2.1
azure-iot-hub>=2.2.1
azure-iot-device>=2.1.3